fix Monolithic Structure
